refactor: remove most ops from the base kernel trait
This makes the kernel more customizable as it no longer "assumes" that
all of these operations will be available.

Also, remove the `LimiterOps` trait entirely, this isn't actually an
"op" (i.e., not a user-invokable operation). The gas ops will get moved
around in a future commit as many of them aren't really "ops" either.